edit recipe

+++ src/main/java/pl/coderslab/repository/DayNameRepository.java
@@ -3,5 +3,5 @@ package pl.coderslab.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pl.coderslab.model.DayName;
 
-public interface DayNameRepository extends JpaRepository<DayName, Integer> {
+public interface DayNameRepository extends JpaRepository<DayName, Long> {
 }
